pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "EstebanDrumshzz",email = "98365132+EstebanDrumshzz@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "langsmith (>=0.3.45,<0.4.0)",
    "langchain-tavily (>=0.2.3,<0.3.0)",
    "requests (>=2.32,<3.0)",
    "grandalf (>=0.8,<0.9)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
